Разбил валидатор на подклассы
