+++ mist_sema/Validators/IValidator.cs
@@ -1,0 +1,9 @@
+﻿using mist_sema.DataClasses;
+
+namespace mist_sema.Validators
+{
+    public interface IValidator
+    {
+        ValidationResult Validate(ComputerConfiguration computerConfiguration);
+    }
+}
